+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+# Mon Dec 27 2021 - version 0.15.0
 ## Ruby server
 - Improved compabitility with Ruby >= 3.0
 - Input parameter options are given as keyword arguments